Rewrite NumberField and NumberInput
